Create README.md
This project is a Windows Forms application in C# that allows to record audio from the user's microphone and save the audio file in WAV format. The application has a graphical interface that includes a button to start recording and a progress bar indicating the remaining recording time. At the end of the recording, a pop-up message is displayed indicating the location of the saved file. The application window is automatically centered on the screen when launched.